requirements.txt
openai>=1.14.3
python-dotenv
